backend/requirements.txt
flask
flask_cors
sqlalchemy
pandas
python-dotenv
psycopg2